Test for React <App /> component
