+++ app/controllers/application_controller.rb
@@ -19,8 +19,4 @@ class ApplicationController < ActionController::Base
     devise_parameter_sanitizer.permit :sign_up, keys: added_attrs
     devise_parameter_sanitizer.permit :account_update, keys: added_attrs
   end
-
-  def after_sign_in_path_for(resource)
-    user_path(resource)
-  end
 end
