+++ app/controllers/static_pages_controller.rb
@@ -1,7 +1,4 @@
 class StaticPagesController < ApplicationController
   def index
-    if user_signed_in?
-      redirect_to user_path(current_user)
-    end
   end
 end
