--- app/controllers/users_controller.rb
@@ -1,6 +1,0 @@
-class UsersController < ApplicationController
-  before_action :authenticate_user!
-
-  def show
-  end
-end
