+++ app/views/static_pages/dashboard.html.erb
