--- app/views/users/index.html.erb
@@ -1,1 +1,0 @@
-
