+++ config/routes.rb
@@ -4,9 +4,18 @@ Rails.application.routes.draw do
   #   root to: "devise/registrations#edit"
   # end
 
-  root 'static_pages#index'
+  # root 'static_pages#index'
 
-  resources :users, only: [:show]
+  # resources :users, only: [:show]
+
+  devise_scope :user do
+    authenticated :user do
+      root "static_pages#dashboard", as: :authenticated_root
+    end
+    unauthenticated :user do
+      root 'static_pages#index', as: :unauthenticated_root
+    end
+  end
 
   namespace :api do
     namespace :v1 do
